+++ Task_LA_08/Shape.java
@@ -1,0 +1,10 @@
+package Task_LA_08;
+
+public class Shape {
+    public String name;
+    public String color;
+
+    public void displayInfo() {
+        System.out.printf("Name: %s\nColor: %s\n", name, color);
+    }
+}
